+++ mlreco/models/layers/common/vertex_ppn.py
@@ -198,8 +198,8 @@ class VertexPPNLoss(torch.nn.modules.loss._Loss):
 
     @staticmethod
     def pairwise_distances(v1, v2):
-        v1_2 = v1.unsqueeze(1).expand(v1.size(0), v2.size(0), v1.size(1)).double()
-        v2_2 = v2.unsqueeze(0).expand(v1.size(0), v2.size(0), v1.size(1)).double()
+        v1_2 = v1.unsqueeze(1).expand(v1.size(0), v2.size(0), v1.size(1))
+        v2_2 = v2.unsqueeze(0).expand(v1.size(0), v2.size(0), v1.size(1))
         return torch.sqrt(torch.pow(v2_2 - v1_2, 2).sum(2))
