+++ mlreco/post_processing/metrics/ppn_simple.py
@@ -10,8 +10,8 @@ from mlreco.utils.ppn import uresnet_ppn_type_point_selector
 
 def pairwise_distances(v1, v2):
     import torch
-    v1_2 = v1.unsqueeze(1).expand(v1.size(0), v2.size(0), v1.size(1)).double()
-    v2_2 = v2.unsqueeze(0).expand(v1.size(0), v2.size(0), v1.size(1)).double()
+    v1_2 = v1.unsqueeze(1).expand(v1.size(0), v2.size(0), v1.size(1))
+    v2_2 = v2.unsqueeze(0).expand(v1.size(0), v2.size(0), v1.size(1))
     return torch.sqrt(torch.pow(v2_2 - v1_2, 2).sum(2))
 
 @post_processing(['ppn-metrics-gt', 'ppn-metrics-pred'],
